Change to support automated build and push
Rename cron-certbot and cron-checkip to replace - with _, thus
cron_certbot and cron_checkip. Underscore in image name is now used to
indicate "inheritance" from another image.

Create build_all script to build all the Docker images in a directory.
Base images are built first, then the "inherited" or derived imaged,
like those mentioned above, are built in a second tranche. This ensures
that the required base images are available.

Re-work other Dockerfiles to enable build-time decision of target
architecture using the --build-arg ARCH. build-all uses this facility
when building for a foreign architecture.